refactor(#5504): keep timeout as long and use DurationUnit.MILLIS
Address review feedback: pass milliseconds as long to the UI, annotate
the property with @DurationUnit(MILLIS), and document that <= 0 disables
auto-dismiss.

+++ spring-boot-admin-server-ui/src/main/java/de/codecentric/boot/admin/server/ui/config/AdminServerUiAutoConfiguration.java
@@ -104,8 +104,7 @@ public class AdminServerUiAutoConfiguration {
 			.favicon(this.adminUi.getFavicon())
 			.faviconDanger(this.adminUi.getFaviconDanger())
 			.enableToasts(this.adminUi.getEnableToasts())
-			.browserNotificationTimeout(
-					(int) this.adminUi.getBrowserNotificationTimeout().toMillis())
+			.browserNotificationTimeout(this.adminUi.getBrowserNotificationTimeout().toMillis())
 			.hideInstanceUrl(this.adminUi.getHideInstanceUrl())
 			.disableInstanceUrl(this.adminUi.getDisableInstanceUrl())
 			.notificationFilterEnabled(
