+++ spring-boot-admin-server-ui/src/main/java/de/codecentric/boot/admin/server/ui/web/UiController.java
@@ -167,7 +167,7 @@ public class UiController {
 
 		private final Boolean enableToasts;
 
-		private final int browserNotificationTimeout;
+		private final long browserNotificationTimeout;
 
 		private final Boolean hideInstanceUrl;
